requirements.txt
iointel
python-dotenv
pyttsx3
matplotlib
transformers          
torch                
